requirements.txt
altgraph==0.17.4
asttokens==2.4.1
attrs==23.2.0
auto-py-to-exe==2.44.0
beautifulsoup4==4.12.3
bottle==0.12.25
bottle-websocket==0.2.9
bs4==0.0.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
Eel==0.16.0
executing==2.0.1
export==0.2.1
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
outcome==1.3.0.post0
packaging==24.1
parso==0.8.4
pefile==2023.2.7
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
pyftpdlib==1.5.9
Pygments==2.18.0
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==26.0.3
requests==2.32.3
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
unzip==1.0.0
urllib3==2.2.1
wcwidth==0.2.13
whichcraft==0.6.1
wsproto==1.2.0
zope.event==5.0
zope.interface==6.4.post2
